+++ tests/integration/cases/record_basic/Java_heterogeneous_strategy_record@jdk_16.java
@@ -1,0 +1,16 @@
+import java.util.Map;
+record Record0(int id, String description, boolean is_done, int[] blocks) {}
+class Main {
+    public static void main() {
+var my_data = new Record0(
+    1,
+    "She said \"hello\", then waved",
+    false,
+    new int[]{
+        1,
+        2,
+        3
+    }
+);
+    }
+}