+++ tests/integration/cases/record_nested_record/Java_heterogeneous_strategy_record@jdk_16.java
@@ -1,0 +1,14 @@
+import java.util.Map;
+record Record1(String name, int age) {}
+record Record0(int id, Record1 owner) {}
+class Main {
+    public static void main() {
+var my_data = new Record0(
+    1,
+    new Record1(
+        "Alice",
+        30
+    )
+);
+    }
+}